+++ tests/ounit_tests/ounit_tests_main.ml
@@ -3,7 +3,6 @@ let suites =
     [
       Ounit_vec_test.suites;
       Ounit_json_tests.suites;
-      Ounit_ext_json_tests.suites;
       Ounit_array_tests.suites;
       Ounit_scc_tests.suites;
       Ounit_list_test.suites;