+++ CHANGELOG.md
@@ -40,6 +40,8 @@
 
 #### :house: Internal
 
+- Move `rescript.json` reading out of `bsc` into rewatch; remove the custom OCaml JSON parser. https://github.com/rescript-lang/rescript/pull/8365
+
 # 13.0.0-alpha.3
 
 #### :boom: Breaking Change